+++ next.config.js
@@ -7,8 +7,9 @@ module.exports = withImages({
 });
 
 const nextConfig = {
-  reactStrictMode: true,
-  // swcMinify: true,
+  // reactStrictMode: true,
+  reactStrictMode: false,
+  swcMinify: true,
 
   webpack(config) {
     config.module.rules.push({
